LayoutTests/fast/css/getComputedStyle/getComputedStyle-resolved-values-expected.txt

Test calling getPropertyValue on computed styles that return "resolved" values.
See http://dev.w3.org/csswg/cssom/#resolved-values for more information.

Resolved values for element "fixed" with display "block":
line-height: 30px
width: 150px
height: 100px
margin-top: 15px
margin-right: 30px
margin-bottom: 20px
margin-left: 10px
padding-top: 20px
padding-right: 20px
padding-bottom: 20px
padding-left: 20px

Resolved values for element "fixed" with display "inline":
line-height: 30px
width: auto
height: auto
margin-top: 15px
margin-right: 30px
margin-bottom: 20px
margin-left: 10px
padding-top: 20px
padding-right: 20px
padding-bottom: 20px
padding-left: 20px

Resolved values for element "fixed" with display "inline-block":
line-height: 30px
width: 150px
height: 100px
margin-top: 15px
margin-right: 30px
margin-bottom: 20px
margin-left: 10px
padding-top: 20px
padding-right: 20px
padding-bottom: 20px
padding-left: 20px

Resolved values for element "fixed" with display "none":
line-height: 30px
width: 150px
height: 100px
margin-top: 15px
margin-right: 30px
margin-bottom: 20px
margin-left: 10px
padding-top: 20px
padding-right: 20px
padding-bottom: 20px
padding-left: 20px


Resolved values for element "percents" with display "block":
line-height: 36px
width: 250px
height: 75px
margin-top: 50px
margin-right: 25px
margin-bottom: 50px
margin-left: 100px
padding-top: 25px
padding-right: 25px
padding-bottom: 25px
padding-left: 25px

Resolved values for element "percents" with display "inline":
line-height: 36px
width: auto
height: auto
margin-top: 10%
margin-right: 5%
margin-bottom: 10%
margin-left: 20%
padding-top: 5%
padding-right: 5%
padding-bottom: 5%
padding-left: 5%

Resolved values for element "percents" with display "inline-block":
line-height: 36px
width: 250px
height: 75px
margin-top: 50px
margin-right: 25px
margin-bottom: 50px
margin-left: 100px
padding-top: 25px
padding-right: 25px
padding-bottom: 25px
padding-left: 25px

Resolved values for element "percents" with display "none":
line-height: 36px
width: 50%
height: 25%
margin-top: 10%
margin-right: 5%
margin-bottom: 10%
margin-left: 20%
padding-top: 5%
padding-right: 5%
padding-bottom: 5%
padding-left: 5%


Resolved values for element "ems" with display "block":
line-height: 24px
width: 240px
height: 120px
margin-top: 24px
margin-right: 12px
margin-bottom: 36px
margin-left: 48px
padding-top: 18px
padding-right: 18px
padding-bottom: 18px
padding-left: 18px

Resolved values for element "ems" with display "inline":
line-height: 24px
width: auto
height: auto
margin-top: 24px
margin-right: 12px
margin-bottom: 36px
margin-left: 48px
padding-top: 18px
padding-right: 18px
padding-bottom: 18px
padding-left: 18px

Resolved values for element "ems" with display "inline-block":
line-height: 24px
width: 240px
height: 120px
margin-top: 24px
margin-right: 12px
margin-bottom: 36px
margin-left: 48px
padding-top: 18px
padding-right: 18px
padding-bottom: 18px
padding-left: 18px

Resolved values for element "ems" with display "none":
line-height: 24px
width: 240px
height: 120px
margin-top: 24px
margin-right: 12px
margin-bottom: 36px
margin-left: 48px
padding-top: 18px
padding-right: 18px
padding-bottom: 18px
padding-left: 18px


Resolved values for element "auto" with display "block":
line-height: 30px
width: 200px
height: 30px
margin-top: 0px
margin-right: 126px
margin-bottom: 0px
margin-left: 126px
padding-top: 0px
padding-right: 0px
padding-bottom: 0px
padding-left: 0px

Resolved values for element "auto" with display "inline":
line-height: 30px
width: auto
height: auto
margin-top: 0px
margin-right: auto
margin-bottom: 0px
margin-left: auto
padding-top: 0px
padding-right: 0px
padding-bottom: 0px
padding-left: 0px

Resolved values for element "auto" with display "inline-block":
line-height: 30px
width: 200px
height: 30px
margin-top: 0px
margin-right: 0px
margin-bottom: 0px
margin-left: 0px
padding-top: 0px
padding-right: 0px
padding-bottom: 0px
padding-left: 0px

Resolved values for element "auto" with display "none":
line-height: 30px
width: 200px
height: auto
margin-top: 0px
margin-right: auto
margin-bottom: 0px
margin-left: auto
padding-top: 0px
padding-right: 0px
padding-bottom: 0px
padding-left: 0px


Resolved values for element "mixed" with display "block":
line-height: 48px
width: 150px
height: 60px
margin-top: 72px
margin-right: 0px
margin-bottom: 0px
margin-left: 252px
padding-top: 40px
padding-right: 50px
padding-bottom: 24px
padding-left: 0px

Resolved values for element "mixed" with display "inline":
line-height: 48px
width: auto
height: auto
margin-top: 72px
margin-right: 0px
margin-bottom: 0px
margin-left: auto
padding-top: 40px
padding-right: 10%
padding-bottom: 24px
padding-left: 0px

Resolved values for element "mixed" with display "inline-block":
line-height: 48px
width: 150px
height: 60px
margin-top: 72px
margin-right: 0px
margin-bottom: 0px
margin-left: 0px
padding-top: 40px
padding-right: 50px
padding-bottom: 24px
padding-left: 0px

Resolved values for element "mixed" with display "none":
line-height: 48px
width: 30%
height: 20%
margin-top: 72px
margin-right: 0px
margin-bottom: 0px
margin-left: auto
padding-top: 40px
padding-right: 10%
padding-bottom: 24px
padding-left: 0px


Fixed
Percents
EMs
Auto
Mixed
